requirements.txt
requests
PyYAML
websocket-client
pandas
numpy